+++ UmaMusumeAPI/Models/Tables/AnnounceSupportCard.cs
@@ -6,10 +6,10 @@ namespace UmaMusumeAPI.Models.Tables
     public class AnnounceSupportCard
     {
         public int Id { get; set; }
-        public long Type { get; set; }
-        public long CutsPattern { get; set; }
-        public long SupportCardIdA { get; set; }
-        public long SupportCardIdB { get; set; }
-        public long SupportCardIdC { get; set; }
+        public int Type { get; set; }
+        public int CutsPattern { get; set; }
+        public int SupportCardIdA { get; set; }
+        public int SupportCardIdB { get; set; }
+        public int SupportCardIdC { get; set; }
     }
 }
