+++ UmaMusumeAPI/Models/Tables/AudienceImpostorColorSet.cs
@@ -2,8 +2,8 @@
 {
     public class AudienceImpostorColorSet
     {
-        public long ColorGroupId { get; set; }
+        public int ColorGroupId { get; set; }
         public string Color { get; set; }
-        public long Odds { get; set; }
+        public int Odds { get; set; }
     }
 }
