+++ UmaMusumeAPI/Models/Tables/AudienceImpostorLottery.cs
@@ -2,9 +2,9 @@
 {
     public class AudienceImpostorLottery
     {
-        public long GroupId { get; set; }
-        public long Pattern { get; set; }
-        public long AssetId { get; set; }
-        public long Odds { get; set; }
+        public int GroupId { get; set; }
+        public int Pattern { get; set; }
+        public int AssetId { get; set; }
+        public int Odds { get; set; }
     }
 }
