+++ UmaMusumeAPI/Models/Tables/AudioStoryEffect.cs
@@ -9,6 +9,6 @@ namespace UmaMusumeAPI.Models.Tables
         public string EffectName { get; set; }
         public string CueName { get; set; }
         public string CueSheet { get; set; }
-        public long LoopSet { get; set; }
+        public int LoopSet { get; set; }
     }
 }
