+++ UmaMusumeAPI/Models/Tables/AvailableSkillSet.cs
@@ -6,8 +6,8 @@ namespace UmaMusumeAPI.Models.Tables
     public class AvailableSkillSet
     {
         public int Id { get; set; }
-        public long AvailableSkillSetId { get; set; }
-        public long SkillId { get; set; }
-        public long NeedRank { get; set; }
+        public int AvailableSkillSetId { get; set; }
+        public int SkillId { get; set; }
+        public int NeedRank { get; set; }
     }
 }
