+++ UmaMusumeAPI/Models/Tables/CardTalentHintUpgrade.cs
@@ -7,15 +7,15 @@ namespace UmaMusumeAPI.Models.Tables
     {
         public int Id { get; set; }
         public int Rarity { get; set; }
-        public long TalentLevel { get; set; }
-        public long ItemCategory1 { get; set; }
-        public long ItemId1 { get; set; }
-        public long ItemDispOrder1 { get; set; }
-        public long ItemNum1 { get; set; }
-        public long ItemCategory2 { get; set; }
-        public long ItemId2 { get; set; }
-        public long ItemDispOrder2 { get; set; }
-        public long ItemNum2 { get; set; }
-        public long MoneyNum { get; set; }
+        public int TalentLevel { get; set; }
+        public int ItemCategory1 { get; set; }
+        public int ItemId1 { get; set; }
+        public int ItemDispOrder1 { get; set; }
+        public int ItemNum1 { get; set; }
+        public int ItemCategory2 { get; set; }
+        public int ItemId2 { get; set; }
+        public int ItemDispOrder2 { get; set; }
+        public int ItemNum2 { get; set; }
+        public int MoneyNum { get; set; }
     }
 }
