+++ UmaMusumeAPI/Models/Tables/ChampionsEvaluationRate.cs
@@ -6,8 +6,8 @@ namespace UmaMusumeAPI.Models.Tables
     public class ChampionsEvaluationRate
     {
         public int Id { get; set; }
-        public long ProperType { get; set; }
-        public long ProperRank { get; set; }
-        public long Rate { get; set; }
+        public int ProperType { get; set; }
+        public int ProperRank { get; set; }
+        public int Rate { get; set; }
     }
 }
