+++ UmaMusumeAPI/Models/Tables/ChampionsNewsTitle.cs
@@ -6,10 +6,10 @@ namespace UmaMusumeAPI.Models.Tables
     public class ChampionsNewsTitle
     {
         public int Id { get; set; }
-        public long RoundId { get; set; }
-        public long WinPercentType { get; set; }
-        public long ResourceId { get; set; }
-        public long Title { get; set; }
-        public long SubTitle { get; set; }
+        public int RoundId { get; set; }
+        public int WinPercentType { get; set; }
+        public int ResourceId { get; set; }
+        public int Title { get; set; }
+        public int SubTitle { get; set; }
     }
 }
