+++ UmaMusumeAPI/Models/Tables/ChampionsRewardRate.cs
@@ -6,13 +6,13 @@ namespace UmaMusumeAPI.Models.Tables
     public class ChampionsRewardRate
     {
         public int Id { get; set; }
-        public long ChampionsId { get; set; }
-        public long LeagueType { get; set; }
-        public long RoundId { get; set; }
-        public long WinCount { get; set; }
-        public long Ranking { get; set; }
-        public long Rate { get; set; }
-        public long RewardSetId { get; set; }
-        public long BoxGrade { get; set; }
+        public int ChampionsId { get; set; }
+        public int LeagueType { get; set; }
+        public int RoundId { get; set; }
+        public int WinCount { get; set; }
+        public int Ranking { get; set; }
+        public int Rate { get; set; }
+        public int RewardSetId { get; set; }
+        public int BoxGrade { get; set; }
     }
 }
