+++ UmaMusumeAPI/Models/Tables/ChampionsRoundDetail.cs
@@ -6,15 +6,15 @@ namespace UmaMusumeAPI.Models.Tables
     public class ChampionsRoundDetail
     {
         public int Id { get; set; }
-        public long ChampionsId { get; set; }
-        public long LeagueType { get; set; }
-        public long RoundId { get; set; }
-        public long Round { get; set; }
-        public long Tier { get; set; }
-        public long RoundNumber { get; set; }
-        public long BreakthroughNumber1 { get; set; }
-        public long BreakthroughNumber2 { get; set; }
-        public long EntryNumber { get; set; }
-        public long FreeEntryNumber { get; set; }
+        public int ChampionsId { get; set; }
+        public int LeagueType { get; set; }
+        public int RoundId { get; set; }
+        public int Round { get; set; }
+        public int Tier { get; set; }
+        public int RoundNumber { get; set; }
+        public int BreakthroughNumber1 { get; set; }
+        public int BreakthroughNumber2 { get; set; }
+        public int EntryNumber { get; set; }
+        public int FreeEntryNumber { get; set; }
     }
 }
