+++ UmaMusumeAPI/Models/Tables/CharaCategoryMotion.cs
@@ -3,11 +3,11 @@
     public class CharaCategoryMotion
     {
         public int CharaId { get; set; }
-        public long StandbyMotion1 { get; set; }
-        public long StandbyMotion2 { get; set; }
-        public long StandbyMotion3 { get; set; }
-        public long StandbyMotion4 { get; set; }
-        public long StandbyMotion5 { get; set; }
-        public long StandbyMotion6 { get; set; }
+        public int StandbyMotion1 { get; set; }
+        public int StandbyMotion2 { get; set; }
+        public int StandbyMotion3 { get; set; }
+        public int StandbyMotion4 { get; set; }
+        public int StandbyMotion5 { get; set; }
+        public int StandbyMotion6 { get; set; }
     }
 }
