+++ UmaMusumeAPI/Models/Tables/CharaMotionAct.cs
@@ -6,8 +6,8 @@ namespace UmaMusumeAPI.Models.Tables
     public class CharaMotionAct
     {
         public int Id { get; set; }
-        public long CharaId { get; set; }
-        public long TargetMotion { get; set; }
+        public int CharaId { get; set; }
+        public int TargetMotion { get; set; }
         public string CommandName { get; set; }
     }
 }
