+++ UmaMusumeAPI/Models/Tables/CharaType.cs
@@ -6,10 +6,10 @@ namespace UmaMusumeAPI.Models.Tables
     public class CharaType
     {
         public string Id { get; set; }
-        public long CharaId { get; set; }
-        public long TargetScene { get; set; }
-        public long TargetCut { get; set; }
-        public long TargetType { get; set; }
-        public long Value { get; set; }
+        public int CharaId { get; set; }
+        public int TargetScene { get; set; }
+        public int TargetCut { get; set; }
+        public int TargetType { get; set; }
+        public int Value { get; set; }
     }
 }
